+++ Sources/PathKit.swift
@@ -472,7 +472,7 @@ extension Path {
     Path.current = self
     defer { Path.current = previous }
     try closure()
-    }
+  }
 }
 
 
@@ -604,7 +604,9 @@ extension Path {
 extension Path {
   public static func glob(_ pattern: String) -> [Path] {
     var gt = glob_t()
-    let cPattern = strdup(pattern)
+    guard let cPattern = strdup(pattern) else {
+      fatalError("strdup returned null: Likely out of memory")
+    }
     defer {
       globfree(&gt)
       free(cPattern)
@@ -632,6 +634,18 @@ extension Path {
 
   public func glob(_ pattern: String) -> [Path] {
     return Path.glob((self + pattern).description)
+  }
+
+  public func match(_ pattern: String) -> Bool {
+    guard let cPattern = strdup(pattern),
+          let cPath = strdup(path) else {
+      fatalError("strdup returned null: Likely out of memory")
+    }
+    defer {
+      free(cPattern)
+      free(cPath)
+    }
+    return fnmatch(cPattern, cPath, 0) == 0
   }
 }