+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # PathKit Changelog
 
+## 1.0.1 (2021-09-22)
+
+### Enhancements
+
+- Adds support for Xcode 13.
+
+- Path's can now be tested against a glob-like pattern using
+  `Path(...).match(pattern)`.
+
 ## 1.0.0 (2019-03-27)
 
 ### Enhancements